# mst: optimize deserializeNodeData allocs, isValidMstKey CPU