@bradfitz 2023-05-02
@Jacob2161 sent me some pprof snapshots. That regexp was using like 3.5% of your CPU.

And GC is ~25% of CPU, which means y'all are allocating a bunch, which the heap pprof confirms, and showed a fair bit in deserializeNodeData (but mostly everywhere). The goal of this PR was really removing the regexp. The deserializeNodeData stuff happened by accident when I was looking at the callers.