@bnewbold 2023-05-02
MST "key" (aka "path" in repo-level nomenclature) changes: sort of wistful for regex, but of course fine. Cross-language test coverage for this behavior exists at a higher-level, via `TestAllowedKeys`.

The `isValidMstKey` function does correctly reject empty key (zero-length string / zero-length bytes), though the way we wrote it doesn't call that out very explicitly. This PR isn't touching those lines so no change needed here.

deserializeNodeData changes: yay! LGTM