@bradfitz 2023-05-02
> sort of wistful for regex, but of course fine

I feel ya. I spent a decade with Perl and love me some regexps. Go's implementation has gotten better over time but it's still not super fast. There are a bunch of fast paths it could have but doesn't.