@bnewbold 2023-05-02
A win with regex is that we can just copy the string around between different language implementations.

A loss is that, actually, you can't always do that, there are a bunch of dialects and corner-cases. Even beyond semantics/"correctness", different langs/impl will have different complexity explosions and denial-of-service vectors. oh well!