@bradfitz 2023-05-02
I suppose I could've at least moved the literal regexp into a test and fuzzed and/or test vector'ed that the custom implementation matches the regexp.