@bradfitz 2023-05-03
I sent https://github.com/bluesky-social/indigo/pull/155 as a peace offering for wistful sadness.